--- saga/Program (2).cs
@@ -1,10 +1,0 @@
-﻿namespace CourseApp
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            Game.Play();
-        }
-    }
-}
